qlog: make all streamer event additions atomic
When we switched from JSON to JSON-SEQ, we modified the qlog serializer but didn't change it's behaviour wrt QUIC Packet events (rx/tx). The old behaviour TL;DR was:

1. qlog streamer user passes streamer a packet event
2. Streamer writes out packet header information. Then it locks itself into frame writing mode.
2. While locked, no other events cand be written.
3. qlog streamer user passes 0 or more frames.
4. qlog streamer user finalizes frames write. Streamer unlocks itself.

There are a few downsides to that approach:

* Locking prevents other events being written. For example, while processing frames, any additional events would be ignored by the serializer. While callers could buffer events themselves and log them later, it adds complication and we decided never to do that inside quiche.
* Related to the above. We had several places where processing might shortcircuit or terminate. We ended up peppering quiche with calls to finalize frame writing.
* It's possible to produce an incomplete log line if the connection is destroyed before.
* It's less efficient to write a partial JSON object out.

The changes in the PR make packet event logging atomic. The qlog streamer is simplified by removing all locking and frame-specific logic. As a consequence, the quiche library requires updates in the packet send() and receive() function. Rather than serialize individual frames as they are added or read from a packet, we buffer the qlog representation of a frame, and then later pass that in a single call to the streamer.