lib: add qlogging of connection close
There are situations where connections are closed but no CONNECTION_CLOSE
frames were exchanged, for example handshake failures or idle timeouts.

This change adds the connectivity::connection_closed event to the mark_closed()
method, just before we blow away the qlog serializer.

I've chosen to have it log an event even if we did send or receive
CONNECTION_CLOSE frames, which is a minor duplication. My rationale is
that in future we might want to have filters that would omit detailed
frame information while key connection state changes are interesting.