qlog: add optional pretty-printing for JSON-SEQ streaming
One of the features of JSON-SEQ is that it can support newlines embedded
in records. Prior to this commit, the qlog streamer only supported a
"condensed mode" of writing events on a single JSON line, which can affect
the usability in some circumstances.

This change adds support for discretionary logging of each individual event
using condensed or pretty formats. A new set of "_pretty" suffixed methods
are provided to allow callers to select the format, without breaking
backwards capability. This design also mimics serde's approach.